restructure hudi-hive-sync tests
